### 2023-01-14 #5077 @editorialbot
[PRE REVIEW]: TextDescriptives: A Python package for calculating a large variety of metrics from text
**Submitting author:** @HLasse (<a href="http://orcid.org/0000-0003-1113-4779">Lasse Hansen</a>)
**Repository:** https://github.com/HLasse/TextDescriptives
**Branch with paper.md** (empty if default branch):
**Version:** v2.1.0
