+++ src/textdescriptives/components/quality_data_classes.py
@@ -237,7 +237,7 @@ class QualityOutput(BaseModel):
 
     def __repr_str__(self, join_str: str) -> str:
         return join_str.join(
-            repr(v) if a is None else f"{a}={v!r}"
+            repr(v) if a is None else f"\n\t{a}={v!r}"
             for a, v in [
                 ("passed", self.passed),
             ]
