### 2023-01-16 50b19ae
2.2.0
Automatically generated by python-semantic-release
