+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 <!--next-version-placeholder-->
 
+## v2.2.0 (2023-01-16)
+### Feature
+* Added QualityOutput ([`c0fb63c`](https://github.com/HLasse/TextDescriptives/commit/c0fb63c671ed1e4ccbe75afa4fb3301104a1ad0e))
+* Updated way that that quality thresholds is set ([`f799186`](https://github.com/HLasse/TextDescriptives/commit/f7991864ba62a4dc0a36f42f2a13ad3079909a02))
+
+### Documentation
+* Removed multiprocessing from pipes ([`c224580`](https://github.com/HLasse/TextDescriptives/commit/c22458097d7ca97679da748e63be6ad1523c9a41))
+* Changed print of quality ([`deb148b`](https://github.com/HLasse/TextDescriptives/commit/deb148bcbe9014c5285fe5a062320f7541b1c1d5))
+* Fixed multiprocessing in tutorial ([`4ddebdf`](https://github.com/HLasse/TextDescriptives/commit/4ddebdf3be5124b92cb4bafe5b69fe1632143aa6))
+* Updated docs with changes to the API ([`580bea1`](https://github.com/HLasse/TextDescriptives/commit/580bea199e91cb95e1876a0e4b7ea0ca5439129d))
+* Updated tutorial ([`57d0054`](https://github.com/HLasse/TextDescriptives/commit/57d00543c1ce0525fee6570b8bb843bc62295bf7))
+* Added quality tutorial to docs ([`2017fb3`](https://github.com/HLasse/TextDescriptives/commit/2017fb351c9b2157bb59525322f509e8367792a6))
+* Added tutorial ([`fa2d65c`](https://github.com/HLasse/TextDescriptives/commit/fa2d65c1114cd6ae03d4638a44e7fe0c525aa80f))
+* Added new tutorial for quality filtering ([`3697e59`](https://github.com/HLasse/TextDescriptives/commit/3697e596cc397c6b22be04f9841f2d55b4721903))
+* Fix typo in coherence.rst ([`be83cf3`](https://github.com/HLasse/TextDescriptives/commit/be83cf354039179ef4d7d25ce4d99e237fc644be))
+* Minor mismention ([`f5155c2`](https://github.com/HLasse/TextDescriptives/commit/f5155c28b5ccced6a334feb1b25f7ae3d4d91854))
+
 ## v2.1.0 (2023-01-06)
 ### Feature
 * Wrapper function ([`fb33e19`](https://github.com/HLasse/TextDescriptives/commit/fb33e19e09b2551a049dccf42722d67f5c17199a))
