+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "textdescriptives"
-version = "2.1.0"
+version = "2.2.0"
 description = "A library for calculating a variety of features from text using spaCy"
 authors = [{name = "Lasse Hansen", email = "lasseh0310@gmail.com"},
              {name = "Kenneth Enevoldsen"}]
