### 2023-01-17 332ba8f
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/charliermarsh/ruff-pre-commit: v0.0.216 → v0.0.223](https://github.com/charliermarsh/ruff-pre-commit/compare/v0.0.216...v0.0.223)
