+++ .pre-commit-config.yaml
@@ -32,7 +32,7 @@ repos:
         args: [--line-length, "88"]
 
   - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: v0.0.216
+    rev: v0.0.223
     hooks:
       - id: ruff
 
