+++ docs/tutorials/requirements.txt
@@ -2,4 +2,4 @@
 # e.g. links to wheels which is not allowed in pyproject.toml on pypi
 
 # spacy pipeline
-https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0.tar.gz
+https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0.tar.gz
