### 2023-01-18 0d39b2e
docs: Added update timer and fail on timeout
