### 2023-01-19 a1177e5
feat: add out of vocabulary ratio to quality component
