+++ pyproject.toml
@@ -36,7 +36,6 @@ dependencies = [
     "spacy>=3.1.0,<3.5.0",
     "numpy>=1.20.0,<1.25.0",
     "pandas>=1.0.0,<1.6.0",
-    "pyphen>=0.11.0,<0.12.0",
     "ftfy>=6.0.3,<6.1.0",
 ]
 
@@ -48,6 +47,9 @@ repository = "https://github.com/HLasse/textdescriptives"
 documentation = "https://hlasse.github.io/TextDescriptives/"
 
 [project.optional-dependencies]
+all = [
+    "pyphen>=0.11.0,<0.12.0",
+]
 style = [
     "black==22.8.0",
     "pre-commit==2.20.0",
