### 2023-01-23 c709f9d
Apply suggestions from code review
Co-authored-by: Lasse Hansen <lasseh0310@gmail.com>
