### 2023-01-24 fd7b20d
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/PyCQA/docformatter: v1.5.1 → v1.6.0.rc1](https://github.com/PyCQA/docformatter/compare/v1.5.1...v1.6.0.rc1)
- [github.com/charliermarsh/ruff-pre-commit: v0.0.223 → v0.0.231](https://github.com/charliermarsh/ruff-pre-commit/compare/v0.0.223...v0.0.231)
