+++ src/textdescriptives/components/utils.py
@@ -29,7 +29,6 @@ def n_tokens(doc: Union[Doc, Span]):
 
 def n_syllables(doc: Doc):
     """Return number of syllables per token."""
-
     dic = Pyphen(lang=doc.lang_)
 
     def count_syl(token: Token):
