+++ tests/test_quality.py
@@ -195,7 +195,6 @@ def test_quality_component(nlp: spacy.Language):
 
 def test_quality_component_with_config(nlp: spacy.Language):
     """Test the quality component with config."""
-
     quality_thresholds = td.QualityThresholds(
         n_stop_words=(3, None),
         alpha_ratio=(None, 0.8),
