+++ src/textdescriptives/components/information_theory.py
@@ -59,7 +59,6 @@ def perplexity_getter(doc: Union[Doc, Span]) -> float:
 
 def per_word_perplexity_getter(doc: Union[Doc, Span]) -> float:
     """Calculates the per word perplexity of a document."""
-
     if hasattr(doc._, "perplexity"):
         perplexity = doc._.perplexity
     else:
