### 2023-02-01 6a57fb9
:arrow_up: Update sphinxext-opengraph requirement
Updates the requirements on [sphinxext-opengraph](https://github.com/wpilibsuite/sphinxext-opengraph) to permit the latest version.
- [Release notes](https://github.com/wpilibsuite/sphinxext-opengraph/releases)
- [Commits](https://github.com/wpilibsuite/sphinxext-opengraph/compare/v0.7.3...v0.7.5)

---
updated-dependencies:
- dependency-name: sphinxext-opengraph
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
