### 2023-02-01 7d5de24
:arrow_up: Bump ruff from 0.0.191 to 0.0.239
Bumps [ruff](https://github.com/charliermarsh/ruff) from 0.0.191 to 0.0.239.
- [Release notes](https://github.com/charliermarsh/ruff/releases)
- [Changelog](https://github.com/charliermarsh/ruff/blob/main/BREAKING_CHANGES.md)
- [Commits](https://github.com/charliermarsh/ruff/compare/v0.0.191...v0.0.239)

---
updated-dependencies:
- dependency-name: ruff
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
