+++ pyproject.toml
@@ -51,7 +51,7 @@ documentation = "https://hlasse.github.io/TextDescriptives/"
 style = [
     "black==22.8.0",
     "pre-commit==2.20.0",
-    "ruff==0.0.191",
+    "ruff==0.0.239",
     "mypy==0.991"
 ]
 tests = [
