+++ .github/workflows/dependabot_automerge.yml
@@ -15,15 +15,16 @@ jobs:
     if: ${{ github.actor == 'dependabot[bot]' || github.actor == 'pre-commit-ci[bot]' }}
 
     steps:
-      # Checkout action is required for token to persist
-      - name: Enable auto-merge for Dependabot PRs
-        run: gh pr merge --auto --merge "$PR_URL" # Use Github CLI to merge automatically the PR
-        env:
-          PR_URL: ${{github.event.pull_request.html_url}}
-          GH_TOKEN: ${{ secrets.DEPENDABOT_AUTOMERGE_TOKEN }}
-
       - name: Auto approve dependabot PRs
         if: ${{ github.actor == 'dependabot[bot]' || github.actor == 'pre-commit-ci[bot]' }}
         uses: hmarr/auto-approve-action@v3.1.0
         with:
           github-token: ${{ secrets.DEPENDABOT_AUTOMERGE_TOKEN }}
+      # Checkout action is required for token to persist
+      - name: Enable auto-merge for Dependabot PRs
+        run: gh pr merge --auto --merge "$PR_URL" # Use Github CLI to merge automatically the PR
+        env:
+          PR_URL: ${{github.event.pull_request.html_url}}
+          GITHUB_TOKEN: ${{ secrets.DEPENDABOT_AUTOMERGE_TOKEN }}
+
+
