### 2023-02-08 729836f
fix: change auto-approve event
Following https://github.com/hmarr/auto-approve-action
