+++ .github/workflows/dependabot_automerge.yml
@@ -2,7 +2,7 @@
 
 name: automerge-dependabot-prs
 
-on: pull_request
+on: pull_request_target
 
 permissions:
   contents: write
