+++ pyproject.toml
@@ -59,7 +59,7 @@ tests = [
     "pytest-cov>=3.0.0,<3.0.1",
 ]
 docs = [
-    "sphinx>=5.3.0,<5.4.0",
+    "sphinx>=5.3.0,<6.2.0",
     "furo==2022.12.7",
     "sphinx-copybutton>=0.5.1,<0.5.2",
     "sphinxext-opengraph>=0.7.3,<0.7.4",
