### 2023-02-09 #5077 @RichardLitt
I may be able to, but I am AFK until Tuesday.

On Thursday, February 9, 2023, Fabian Scheipl ***@***.***>
wrote:

> hi @paragkulkarni11 <https://github.com/paragkulkarni11>, @gweissman
> <https://github.com/gweissman>, @mbod <https://github.com/mbod>
>
> would any of you be willing to review this submission for JOSS?
>
> We carry out our checklist-driven reviews here in GitHub issues and follow
> these guidelines: https://joss.readthedocs.io/
> en/latest/review_criteria.html
>
