### 2023-02-14 4f37b6b
ci: make docformatter and black compatible
