+++ .pre-commit-config.yaml
@@ -23,7 +23,15 @@ repos:
     rev: v1.6.0.rc1
     hooks:
       - id: docformatter
-        args: [--in-place]
+        args:
+          [
+            --in-place,
+            --pre-summary-space,
+            --wrap-descriptions,
+            "88",
+            --wrap-summaries,
+            "88",
+          ]
 
   - repo: https://github.com/psf/black
     rev: 23.1.0
