+++ tests/books.py
@@ -1,4 +1,4 @@
-"""These books and several tests are borrowed from
+""" These books and several tests are borrowed from
 https://github.com/mholtzscher/spacy_readability."""
 
 oliver_twist = """Among other public buildings in a certain town, which for many reasons
