+++ src/textdescriptives/components/readability.py
@@ -14,7 +14,8 @@ from .utils import filter_tokens
 class Readability:
     """spaCy v.3.0 component for adding readability metrics to `Doc` objects.
 
-    Extracts metrics and returns them as a dictionary as the ._.readability attribute.
+    Extracts metrics and returns them as a dictionary as the ._.readability
+    attribute.
     """
 
     def __init__(self, nlp: Language):
@@ -23,8 +24,8 @@ class Readability:
             Doc.set_extension("readability", getter=self.readability)
 
     def _flesch_reading_ease(self, doc: Doc):
-        """Calculate the Flesch Reading Ease score for a document. The equation for the
-        Flesch Reading Ease score is:
+        """Calculate the Flesch Reading Ease score for a document. The equation
+        for the Flesch Reading Ease score is:
 
         206.835 - (1.015 X avg sent len) - (84.6 * avg_syl_per_word)
 
@@ -38,8 +39,8 @@ class Readability:
         return score
 
     def _flesch_kincaid_grade(self, doc: Doc):
-        """Calculate the Flesch-Kincaid grade of the document. The equation for the
-        Flesch-Kincaid grade is:
+        """Calculate the Flesch-Kincaid grade of the document. The equation for
+        the Flesch-Kincaid grade is:
 
         0.39 * (avg sent len) + 11.8 * (avg_syl_per_word) - 15.59
         """
@@ -51,7 +52,8 @@ class Readability:
         return score
 
     def _smog(self, doc: Doc, n_hard_words: int):
-        """Calculate the SMOG index of the document. The equation for the SMOG index is:
+        """Calculate the SMOG index of the document. The equation for the SMOG index
+        is:
 
         1.043( sqrt(30 * (hard words /n sentences)) + 3.1291
 
@@ -65,8 +67,8 @@ class Readability:
         return np.nan
 
     def _gunning_fog(self, doc, n_hard_words: int):
-        """Calculates the Gunning Fog index of the document. The equation for the
-        Gunning Fog index is:
+        """Calculates the Gunning Fog index of the document. The equation for
+        the Gunning Fog index is:
 
         Grade level = 0.4 * ((avg_sentence_length) + (percentage hard words))
 
@@ -80,8 +82,8 @@ class Readability:
         return 0.4 * (avg_sent_len + percent_hard_words)
 
     def _automated_readability_index(self, doc: Doc):
-        """Calculates the Automated Readability Index of the document. The equation for
-        the Automated Readability Index is:
+        """Calculates the Automated Readability Index of the document. The
+        equation for the Automated Readability Index is:
 
         4.71 * (n_chars / n_words) + 0.5 * (n_words / n_sentences) - 21.43
 
@@ -97,8 +99,8 @@ class Readability:
         return score
 
     def _coleman_liau_index(self, doc: Doc):
-        """Calculates the Colmean-Liau index of the document. The equation for the
-        Coleman-Liau index is:
+        """Calculates the Colmean-Liau index of the document. The equation for
+        the Coleman-Liau index is:
 
         score = 0.0588 * avg number of chars pr 100 words - 0.296 * avg num of sents pr
         100 words -15.8
@@ -113,7 +115,8 @@ class Readability:
         return 0.0588 * lengths - 0.296 * s - 15.8
 
     def _lix(self, doc: Doc, long_words: int):
-        """Calculates the LIX index of the document. The equation for the LIX index is:
+        """Calculates the LIX index of the document. The equation for the LIX index
+        is:
 
         (n_words / n_sentences) + (n_words longer than 6 characters * 100) / n_words
         """
@@ -124,7 +127,8 @@ class Readability:
         return doc._.sentence_length["sentence_length_mean"] + percent_long_words
 
[…]
