### 2023-02-14 5ad185d
ci: remove docformatter due to black incompatibility
