+++ paper/paper.bib
@@ -113,6 +113,7 @@
 	volume = {9},
 	pages = {159--191},
 	number = {2},
+	doi = {10.17791/jcs.2008.9.2.159},
 	journaltitle = {Journal of Cognitive Science},
 	author = {Liu, Haitao},
 	date = {2008},
