### 2023-02-15 #5077 @fccoelho
I can do it as well, In case you still need a second reviewer.

best,

Flávio

On Wed, Feb 15, 2023 at 8:32 AM Fabian Scheipl ***@***.***>
wrote:

> hi @fccoelho <https://github.com/fccoelho> @hadware
> <https://github.com/hadware>
> would you be willing to review this submission for JOSS? We carry out our
> checklist-driven reviews here in GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
