### 2023-03-01 #5153 @RichardLitt
See https://github.com/HLasse/TextDescriptives/issues/184 for current status of my checklist.
