### 2023-03-01 #5153 @RichardLitt
> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I would like this to be clearer. @HLasse - can you add a Contributing guide?
