### 2023-03-01 #5153 @fabian-s
@hadware - thank you for starting your review.

For the items on your checklist that unchecked, you are encouraged to submit issues and pull requests on the software repository. When doing so, please mention openjournals/joss-reviews#5153 so that a link is created to this thread (and I can keep an eye on what is happening).
Please also feel free to comment and ask questions on this thread re. the unchecked items (or other points you come across during the review) so that any problems can be resolved quickly and collaboratively with @HLasse.
