### 2023-03-01 254f4e2
Create CODE_OF_CONDUCT.md
