### 2023-03-01 #5153 @KennethEnevoldsen
@HLasse I have created a [PR](https://github.com/HLasse/TextDescriptives/pull/185) to solve the lack of contributing guidelines.
