### 2023-03-01 680b52f
tests: add test for `extract_metric` with `metrics=None`
