### 2023-03-01 #5153 @HLasse
> @HLasse Is there a reason that Ludvig isn't listed as a coauthor? He had some important commits early on, from what I read.

He put in a great effort in refactoring the early version (that was based on Stanford NLP), but has not been active since the switch to spaCy. I asked him if he would like to be coauthor (and help with maintenance) and he accepted. I will add his name to the paper!

> @HLasse I have created a https://github.com/HLasse/TextDescriptives/pull/185 to solve the lack of contributing guidelines.

Thanks, Kenneth! I've added some minor comments to the PR.
