### 2023-03-01 56f5ff4
paper: remove Ludvig from acknowledgements
