### 2023-03-01 HLasse/TextDescriptives#184 @HLasse
Hi Richard, thanks for pointing this out! There was a bug when setting the metrics argument of `extract_metrics` to `None` which is fixed in #186 and will be merged once CI is finished running.
