+++ paper/readme.md
@@ -2,7 +2,7 @@
 
 `paper.md` contains the paper submitted to JOSS
 
-`paper_quarter.qmd` contains the same contant as `paper.md` but with a different yaml preamble that allows it to be converted to a nice arxiv-style pdf. To render to arxiv preprint do the following:
+`paper_quarter.qmd` contains the same content as `paper.md` but with a different yaml preamble that allows it to be converted to a nice arxiv-style pdf. To render to arxiv preprint do the following:
 
 ```
 # install the quarto-arxiv template (https://github.com/mikemahoney218/quarto-arxiv)
