### 2023-03-01 f54a9b6
fix: automatically download supplied spacy model if not on disk
