### 2023-03-01 cdd85b7
docs: update readme to use lg model for proper coherence calculation
