### 2023-03-01 ddfd269
fix: change default verbosity of readability component
