### 2023-03-01 3fd3240
fix: add descriptive_stats pipe no matter the verbosity
