+++ _sources/dependencydistance.rst.txt
@@ -1,7 +1,7 @@
 Dependency Distance
 --------------------
 
-The *dependency_distance* component adds measures of depedency distance to both :code:`Doc`, :code:`Span`, and :code:`Token` objects 
+The *dependency_distance* component adds measures of dependency distance to both :code:`Doc`, :code:`Span`, and :code:`Token` objects 
 under the :code:`._.dependency_distance` attribute.
 Dependency distance can be used a measure of syntactics complexity, the greater the distance, the more complex 
 (`Liu 2008 <https://pdfs.semanticscholar.org/b6b9/cf00698a76d7a1e5ba58baa92d8799366813.pdf>`__, `Oya 2011 <http://www.paaljapan.org/conference2011/ProcNewest2011/pdf/poster/P-13.pdf>`__). 
