### 2023-03-02 187384a
Fix CITATION.cff file name and invalid DOI value
- Change the file name to adhere to the CFF format definition: https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md#general-structure-of-a-citationcff-file
- Fix the DOI value to be the DOI, not the resolver URL: https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md#definitionsdoi
