+++ citation.cff
@@ -1,28 +1,0 @@
-cff-version: 1.2.0
-message: "If you use this software, please cite it as below."
-authors:
-  - family-names: "Hansen"
-    given-names: "Lasse"
-    orcid: "https://orcid.org/0000-0003-1113-4779"
-  - family-names: "Enevoldsen"
-    given-names: "Kenneth"
-    orcid: "https://orcid.org/0000-0001-8733-0966"
-
-title: "TextDescriptives: A Python package for calculating a large variety of statistics from text"
-version: 2.0.10
-date-released: 2020-01-28
-url: "https://github.com/HLasse/TextDescriptives"
-
-preferred-citation:
-  type: article
-  authors:
-    - family-names: "Hansen"
-      given-names: "Lasse"
-      orcid: "https://orcid.org/0000-0003-1113-4779"
-    - family-names: "Enevoldsen"
-      given-names: "Kenneth"
-      orcid: "https://orcid.org/0000-0001-8733-0966"
-  doi: "https://doi.org/10.48550/arxiv.2301.02057"
-  journal: "arXiv"
-  title: "TextDescriptives: A Python package for calculating a large variety of statistics from text"
-  year: 2023
