+++ pyproject.toml
@@ -50,7 +50,7 @@ documentation = "https://hlasse.github.io/TextDescriptives/"
 [project.optional-dependencies]
 style = [
     "black==22.8.0",
-    "pre-commit==3.0.4",
+    "pre-commit==3.1.1",
     "ruff==0.0.253",
     "mypy==1.0.1"
 ]
