### 2023-03-02 ae96a16
citation: add Ludvig to author list
