### 2023-03-03 HLasse/TextDescriptives#184 @HLasse
That's odd. I can't reproduce on a fresh virtual environment with textdescriptives version 2.4.3. Which version are you using?
