### 2023-03-03 b4baed8
:arrow_up: Update datasets requirement
Updates the requirements on [datasets](https://github.com/huggingface/datasets) to permit the latest version.
- [Release notes](https://github.com/huggingface/datasets/releases)
- [Commits](https://github.com/huggingface/datasets/compare/2.8.0...2.10.1)

---
updated-dependencies:
- dependency-name: datasets
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
