+++ pyproject.toml
@@ -71,7 +71,7 @@ tutorials = [
     "jupyter",
     "seaborn",
     "matplotlib",
-    "datasets>=2.8.0,<2.9.0",
+    "datasets>=2.8.0,<2.11.0",
 ]
 
 [project.readme]
