### 2023-03-05 e9138b0
readme: add how to download spacy pipeline
