### 2023-03-09 #5153 @fabian-s
hi @hadware  :wave: -- please reply to / take action re. my previous message:

For the items on your checklist that unchecked, you are encouraged to submit issues and pull requests on the software repository. When doing so, please mention https://github.com/openjournals/joss-reviews/issues/5153 (`https://github.com/openjournals/joss-reviews/issues/5153`) so that a link is created to this thread (and I can keep an eye on what is happening).
Please also comment and ask questions on this thread re. the unchecked items (or other points you come across during the review) so that any problems can be resolved quickly and collaboratively with @HLasse
and, importantly, to let them know *what* the specific issues are that you have identified.
