### 2023-03-15 HLasse/TextDescriptives#199 @HLasse
Hi @RichardLitt. It seems you're trying to use the [test](https://docs.python.org/3/library/test.html#module-test) package for running the test.

We use `pytest` for testing the package. There is a guide to running the tests in the FAQ in the documentation  [here](https://hlasse.github.io/TextDescriptives/faq.html).

To sum up:
Try installing the package using the following lines:

```
pip install -e ."[style,tests]"
pip install -r tests/requirements.txt
````
Then you can run the tests from the root directory using
```
pytest
```
