+++ .pre-commit-config.yaml
@@ -53,7 +53,7 @@ repos:
         #args: [--line-length, "88"]
 
   - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: v0.0.255
+    rev: v0.0.257
     hooks:
       - id: ruff
 
