### 2023-03-28 1e38fff
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/charliermarsh/ruff-pre-commit: v0.0.257 → v0.0.259](https://github.com/charliermarsh/ruff-pre-commit/compare/v0.0.257...v0.0.259)
