+++ pyproject.toml
@@ -52,7 +52,7 @@ style = [
     "black==22.8.0",
     "pre-commit==3.1.1",
     "ruff==0.0.253",
-    "mypy==1.0.1"
+    "mypy==1.1.1"
 ]
 tests = [
     "pytest>=7.1.3,<7.3.0",
