+++ pyproject.toml
@@ -49,7 +49,7 @@ documentation = "https://hlasse.github.io/TextDescriptives/"
 
 [project.optional-dependencies]
 style = [
-    "black==22.8.0",
+    "black==23.3.0",
     "pre-commit==3.1.1",
     "ruff==0.0.253",
     "mypy==1.1.1"
