+++ .pre-commit-config.yaml
@@ -47,13 +47,13 @@ repos:
   #         ]
 
   - repo: https://github.com/psf/black
-    rev: 23.1.0
+    rev: 23.3.0
     hooks:
       - id: black
         #args: [--line-length, "88"]
 
   - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: v0.0.257
+    rev: v0.0.260
     hooks:
       - id: ruff
 
