### 2023-04-17 #5153 @fabian-s
:wave: @linuxscout @timClicks would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html.
Another reviewer has already helped to improve the paper/software, so all we need from you would be a quick confirmation of the JOSS checklist items from you for this.
