### 2023-04-17 #5153 @hadware
> thanks for the nudge. I'll kick hadware off this paper since they are not replying anymore. I'll never understand why people volunteer to review and then disappear a week later and stop responding. sucks.

I'm sorry for being very unresponsive. I wanted to do a thorough review of the code (as I always do for JOSS) but never could find the time to do so. The previous nudge from 1 month ago probably also got lost in the list of notifications and emails I constantly get from github.
