### 2023-04-19 c682f1c
Replaces all pos_stats with pos_proportions
