+++ src/textdescriptives/extractors.py
@@ -39,7 +39,7 @@ def extract_dict(
         docs (Union[Iterable[Doc],  Doc]): An iterable of spaCy Docs or a single Doc
         metrics (Union[list[str], str, None], optional): Which metrics to extract.
                 One or more of ["descriptive_stats", "readability",
-                "dependency_distance", "pos_stats", "information_theory"].
+                "dependency_distance", "pos_proportions", "information_theory"].
                 Defaults to None in which case it will extract metrics for which a
                 pipeline compoenent has been set.
         include_text (bool, optional): Whether to add an entry containing the text.
@@ -95,7 +95,7 @@ def extract_df(
         docs (Union[Iterable[Doc],  Doc]): An iterable of spaCy Docs or a single Doc
         metrics (Union[list[str], str], optional): Which metrics to extract.
                 One or more of ["descriptive_stats", "readability",
-                "dependency_distance", "pos_stats"]. Defaults to None in which
+                "dependency_distance", "pos_proportions"]. Defaults to None in which
                 case it will extract metrics for which a pipeline compoenent has been
                 set.
         include_text (bool, optional): Whether to add a column containing the text.
@@ -125,7 +125,7 @@ def extract_metrics(
             model for the language. Defaults to None.
         metrics (List[str]): Which metrics to extract.
             One or more of ["descriptive_stats", "readability",
-            "dependency_distance", "pos_stats", "coherence", "quality"]. If None,
+            "dependency_distance", "pos_proportions", "coherence", "quality"]. If None,
             will extract all metrics from textdescriptives. Defaults to None.
         spacy_model_size (str, optional): Size of the spacy model to download.
 
