### 2023-04-19 3f747f2
readme: add app as badge and to news
