### 2023-04-19 cb55e23
fixed fail on empty string
