### 2023-04-19 2bed2c1
fix: dep distance test
unsure why this failed. Potentially due to a changed model?
