### 2023-04-19 abfa507
fix: coherence model now handles empty strings as intended
