+++ _sources/tutorials/filter_corpus_using_quality.ipynb.txt
@@ -73,13 +73,13 @@
     "from datasets import load_dataset\n",
     "\n",
     "# stream in the dataset\n",
-    "dataset = load_dataset(\"mc4\", \"en\", streaming = True, split = \"train\")\n",
+    "dataset = load_dataset(\"mc4\", \"en\", streaming=True, split=\"train\")\n",
     "\n",
     "# download the first 1 000\n",
     "dataset = dataset.take(1000)\n",
     "\n",
     "# extract the text\n",
-    "texts = [sample [\"text\"] for sample in dataset]\n"
+    "texts = [sample[\"text\"] for sample in dataset]"
    ]
   },
   {
@@ -98,7 +98,7 @@
    ],
    "source": [
     "# let us look at the first part (400 characters) of the first text\n",
-    "print(texts[0][:400])\n"
+    "print(texts[0][:400])"
    ]
   },
   {
@@ -313,7 +313,9 @@
     }
    ],
    "source": [
-    "print(f\"A total of {len(docs)} texts were processed and {len(filtered_texts)} passed the quality check.\")"
+    "print(\n",
+    "    f\"A total of {len(docs)} texts were processed and {len(filtered_texts)} passed the quality check.\"\n",
+    ")"
    ]
   },
   {
@@ -332,18 +334,20 @@
    "outputs": [],
    "source": [
     "new_thresholds = td.QualityThresholds(\n",
-    "    n_stop_words=(2, None), # at least 2 stop words, no upper bound\n",
-    "    alpha_ratio= (0.7, None),\n",
-    "    mean_word_length= (3, 10), # mean word length between 3 and 10 characters\n",
-    "    doc_length = (10, 100_000),\n",
-    "    symbol_to_word_ratio = {}, # don't filter based on symbol to word ratio.\n",
-    "    proportion_ellipsis = (None, 0.3),\n",
-    "    proportion_bullet_points = (None, 0.8),\n",
-    "    contains = {\"lorem ipsum\": False}, # remove texts which contain the string \"lorem ipsum\"\n",
-    "    duplicate_line_chr_fraction = (None, 0.2),\n",
-    "    duplicate_paragraph_chr_fraction = (None, 0.2),\n",
-    "    duplicate_ngram_chr_fraction = {}, # don't filter based on duplicate n-grams\n",
-    "    top_ngram_chr_fraction = {\"2\": (None, 0.2), \"3\": (None, 0.18), \"4\": (None, 0.16)}\n",
+    "    n_stop_words=(2, None),  # at least 2 stop words, no upper bound\n",
+    "    alpha_ratio=(0.7, None),\n",
+    "    mean_word_length=(3, 10),  # mean word length between 3 and 10 characters\n",
+    "    doc_length=(10, 100_000),\n",
+    "    symbol_to_word_ratio={},  # don't filter based on symbol to word ratio.\n",
+    "    proportion_ellipsis=(None, 0.3),\n",
+    "    proportion_bullet_points=(None, 0.8),\n",
+    "    contains={\n",
+    "        \"lorem ipsum\": False\n",
+    "    },  # remove texts which contain the string \"lorem ipsum\"\n",
+    "    duplicate_line_chr_fraction=(None, 0.2),\n",
+    "    duplicate_paragraph_chr_fraction=(None, 0.2),\n",
+    "    duplicate_ngram_chr_fraction={},  # don't filter based on duplicate n-grams\n",
+    "    top_ngram_chr_fraction={\"2\": (None, 0.2), \"3\": (None, 0.18), \"4\": (None, 0.16)},\n",
     ")\n",
     "\n",
     "# overwrite the existing thresholds\n",
@@ -771,7 +775,9 @@
[…]
