### 2023-04-22 #5153 @linuxscout
Hi @HLasse,

I came across your paper, and I must say that I found it to be an excellent piece of work. Your writing was clear and concise, and the conclusions were very compelling.

I also wanted to express my appreciation for the package "TextDescriptives" that you have developed. It is incredibly useful in many works on text analysis. I was wondering if the package is designed only for English or if it can be used for other languages or non-Latin languages as well. I would be interested in knowing if there are any limitations to using the package with non-English texts.

Best regards,
