### 2023-04-22 #5153 @KennethEnevoldsen
Hi @linuxscout,

I Will just answer here as the co-author (@HLasse feel free to fill out anything I missed).

The package is designed to be applicable in multiple languages as long as it has a spacy pipeline or works well with their multilingual pipeline. Parts of the components (e.g. dependency distance, pos proportion) requires a trained component and are of course limited to languages for which these components exist (SpaCy includes ~23 languages), but additional languages can be supported using e.g. [spacy-stanza](https://pypi.org/project/spacy-stanza/).

Some components such as the quality filter does have some assumptions built into their default parameters (e.g. average word length varies for languages with compounds words such as Danish and German) and if you want to filter text using that component those parameters would have to be tuned.

We have used it for both Danish and English ourselves (potentially Norwegian and Swedish as well), but have not tested out on non-latin scripts (although many of the metrics or underlying concepts have been applied in non-latin languages).
