### 2023-04-24 #5153 @HLasse
Thanks for the excellent feedback and the nice comments!

A tagged release has been made and archived on Zenodo - v2.4.5 (https://zenodo.org/record/7847448)

The doi of the tagged release is: https://doi.org/10.5281/zenodo.7847448
