### 2023-04-24 b7ecced
paper: add information theory component and web app
