### 2023-04-24 #5153 @HLasse
I see that we haven't updated the paper with the new features added in the last month or two. I have uploaded a new version of the paper that adds information on the `information_theory` component and a short paragraph on the new [web app](https://huggingface.co/spaces/HLasse/textdescriptives)
