### 2023-04-24 #5153 @HLasse
I'm happy with the proof! 👍
